requirements.txt
google-generativeai>=0.3.0
langgraph>=0.0.10
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0 